Fixed bug about buying a seat reserved and then freed

+++ book.php
@@ -73,6 +73,7 @@
         global $conn;
         $id = $_REQUEST['cell'];
         $user_logged = $_SESSION['265444_user'];
+        $_SESSION[$id] = 1;
         $conn->autocommit(FALSE);
         $conn->begin_transaction();
         $query = "INSERT INTO booking(SeatId, Status, Username) VALUES (?, ?, ?)
@@ -98,6 +99,7 @@
         global $conn;
         $id = $_REQUEST['cell'];
         $user_logged = $_SESSION['265444_user'];
+        $_SESSION[$id] = 0;
         $conn->autocommit(FALSE);
         $conn->begin_transaction();
         $query = "DELETE FROM booking WHERE SeatId=? AND Username=?";
@@ -119,8 +121,11 @@
 
     function purchaseSeats(){
         global $conn;
+        global $rows, $columns;
         $n_cells = $_REQUEST['ncells'];
         $user_logged = $_SESSION['265444_user'];
+        $continue = 0;
+        $reserved_view_seats = array();
         $conn->autocommit(FALSE);
         $conn->begin_transaction();
         $query = "SELECT COUNT(*) FROM booking WHERE Username=? AND Status='R'";
@@ -134,17 +139,66 @@
             mysqli_stmt_fetch($stmt);
             mysqli_stmt_close($stmt);
             if($n_reserved != $n_cells){
-                $query = "DELETE FROM booking WHERE Username=? AND Status='R'";
-                if ($stmt = mysqli_prepare($conn, $query)) {
-                    mysqli_stmt_bind_param($stmt, "s", $user_logged);
-                    if(!mysqli_stmt_execute($stmt)){
-                        $conn->rollback();
-                        return false;
+                //echo $n_reserved;
+                for($i = 0; $i < $rows; $i++){
+                    for($j = 0; $j < $columns; $j++){
+                        $car = chr(65+$j);
+                        $stringId = $car.($i+1);
+                        if(isset($_SESSION[$stringId])){
+                            if($_SESSION[$stringId] == 1){
+                                 //inserire qui query per vedere se questo 
+                                 //posto è presente nel database oppure se è libero   
+                                $query = "SELECT COUNT(*), Username FROM booking WHERE SeatId = ? FOR UPDATE";
+                                if($stmt = mysqli_prepare($conn, $query)){
+                                    mysqli_stmt_bind_param($stmt, "s", $stringId);
+                                    if(!mysqli_stmt_execute($stmt)){
+                                        $conn->rollback();
+                                        return false;
+                                    }
+                                    mysqli_stmt_bind_result($stmt, $is_there, $is_whose);
+                                    mysqli_stmt_fetch($stmt);
+                                    mysqli_stmt_close($stmt);
+                                    if($is_there != 0 && $is_whose != $user_logged){
+                                        $continue = 1;
+                                    }
+                                    else if($is_there == 0){
+                                        $query = "INSERT INTO booking(SeatId, Status, Username) VALUES (?, ?, ?)";
+                                        $status_new = 'R';
+                                        if ($stmt = mysqli_prepare($conn, $query)) {
+                                            mysqli_stmt_bind_param($stmt, "sss", $stringId, $status_new, $user_logged);
+                                            if(!mysqli_stmt_execute($stmt)){
+                                                $conn->rollback();
+                                                return false;
+                                            }
+                                            mysqli_stmt_close($stmt);
+                                        }
+                                        else{
+                                            $conn->rollback();
+                                            return false;
+                                        }
+                                    }
+                                }
+                                else{
+                                    $conn->rollback();
+                                    return false;
+                                }
+                            }
+                        }
                     }
-                    mysqli_stmt_close($stmt);
                 }
-                $conn->commit();
-                return false;
+                if($continue == 1){
+                    $query = "DELETE FROM booking WHERE Username=? AND Status='R'";
+                    if ($stmt = mysqli_prepare($conn, $query)) {
+                        mysqli_stmt_bind_param($stmt, "s", $user_logged);
+                        if(!mysqli_stmt_execute($stmt)){
+                            $conn->rollback();
+                            return false;
+                        }
+                        mysqli_stmt_close($stmt);
+                    }
+                    $conn->commit();
+                    return false;
+                }
             }
         }
         else{
